--- src/Component/Header/Modal/CalendarDetail.tsx
@@ -1,5 +1,0 @@
-const CalendarDetail = () => {
-	return <div>Calendar</div>;
-};
-
-export default CalendarDetail;
