--- src/Component/Header/Modal/PriceDetail.tsx
@@ -1,5 +1,0 @@
-const PriceDetail = () => {
-	return <div>Price</div>;
-};
-
-export default PriceDetail;
