+++ src/Component/Header/SearchBar/Schedule/Schedule.tsx
@@ -21,7 +21,7 @@ const Schedule = () => {
 			<StyledCheckin>
 				<StyledSearchBarChild onClick={() => checkModal(modalStateCheckin)} name={name}>
 					<div>체크인</div>
-					<div>{checkin.year ? `${checkin.month}월${checkin.date}일` : "날짜 입력"}</div>
+					<div>{checkin.year ? `${checkin.month + 1}월${checkin.date}일` : "날짜 입력"}</div>
 				</StyledSearchBarChild>
 				{isCheckin && (
 					<CancelButton
@@ -37,7 +37,7 @@ const Schedule = () => {
 			<StyledCheckout>
 				<StyledSearchBarChild onClick={() => checkModal(modalStateCheckout)} name={name}>
 					<div>체크아웃</div>
-					<div>{checkout.year ? `${checkout.month}월${checkout.date}일` : "날짜 입력"}</div>
+					<div>{checkout.year ? `${checkout.month + 1}월${checkout.date}일` : "날짜 입력"}</div>
 				</StyledSearchBarChild>
 				{isCheckout && (
 					<CancelButton
